+++ src/Umbraco.Web.UI.Client/src/packages/core/collection/action/index.ts
@@ -2,4 +2,4 @@ import './collection-action-button.element.js';
 
 export { UmbCollectionActionBase } from './collection-action-base.js';
 export { UmbCollectionActionButtonElement as UmbCollectionActionElement } from './collection-action-button.element.js';
-export { UmbPopoverScrollElement } from './create/collection-create-popover-scroll.mixin.js';
+export { UmbPopoverScrollElement } from './popover-scroll.element.js';
