+++ src/Umbraco.Web.UI.Client/src/packages/media/media/collection/action/create-media-collection-action.element.ts
@@ -92,7 +92,7 @@ export class UmbCreateMediaCollectionActionElement extends UmbPopoverScrollEleme
 		return html`
 			<uui-button popovertarget="collection-action-menu-popover" label=${label} color="default" look="outline">
 				${label}
-				<uui-symbol-expand .open=${this._popoverOpen}></uui-symbol-expand>
+				<uui-symbol-expand .open=${this.popoverOpen}></uui-symbol-expand>
 			</uui-button>
 			<uui-popover-container
 				id="collection-action-menu-popover"
