+++ src/Umbraco.Web.UI.Client/src/packages/members/member/collection/action/create-member-collection-action.element.ts
@@ -65,7 +65,7 @@ export class UmbCreateMemberCollectionActionElement extends UmbPopoverScrollElem
 		return html`
 			<uui-button popovertarget="collection-action-menu-popover" label=${label} color="default" look="outline">
 				${label}
-				<uui-symbol-expand .open=${this._popoverOpen}></uui-symbol-expand>
+				<uui-symbol-expand .open=${this.popoverOpen}></uui-symbol-expand>
 			</uui-button>
 			<uui-popover-container
 				id="collection-action-menu-popover"
